path refactoring
