+++ manager/src/Controller/Profile/ShowController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace App\Controller;
+namespace App\Controller\Profile;
 
 
 use App\ReadModel\User\UserFetcher;
